Struct bevy_color::prelude::Hsla

pub struct Hsla {
    pub hue: f32,
    pub saturation: f32,
    pub lightness: f32,
    pub alpha: f32,
}

Color in Hue-Saturation-Lightness (HSL) color space with alpha. Further information on this color model can be found on Wikipedia.

Conversion

Conversion between the various color spaces is achieved using Rust’s native From trait. Because certain color spaces are defined by their transformation to and from another space, these From implementations reflect that set of definitions.

let color = Srgba::rgb(0.5, 0.5, 0.5);

// Using From explicitly
let linear_color = LinearRgba::from(color);

// Using Into
let linear_color: LinearRgba = color.into();

For example, the sRGB space is defined by its relationship with Linear RGB, and HWB by its with sRGB. As such, it is the responsibility of sRGB to provide From implementations for Linear RGB, and HWB for sRGB. To then provide conversion between Linear RGB and HWB directly, HWB is responsible for implementing these conversions, delegating to sRGB as an intermediatory. This ensures that all conversions take the shortest path between any two spaces, and limit the proliferation of domain specific knowledge for each color space to their respective definitions.

Fields
hue: f32

The hue channel. [0.0, 360.0]

saturation: f32

The saturation channel. [0.0, 1.0]

lightness: f32

The lightness channel. [0.0, 1.0]

alpha: f32

The alpha channel. [0.0, 1.0]

Implementations

impl Hsla

pub const fn new(hue: f32, saturation: f32, lightness: f32, alpha: f32) -> Self

Construct a new Hsla color from components.

Arguments

hue - Hue channel. [0.0, 360.0]

saturation - Saturation channel. [0.0, 1.0]

lightness - Lightness channel. [0.0, 1.0]

alpha - Alpha channel. [0.0, 1.0]

pub const fn hsl(hue: f32, saturation: f32, lightness: f32) -> Self

Construct a new Hsla color from (h, s, l) components, with the default alpha (1.0).

Arguments

hue - Hue channel. [0.0, 360.0]

saturation - Saturation channel. [0.0, 1.0]

lightness - Lightness channel. [0.0, 1.0]

pub const fn with_saturation(self, saturation: f32) -> Self

Return a copy of this color with the saturation channel set to the given value.

pub const fn with_lightness(self, lightness: f32) -> Self

Return a copy of this color with the lightness channel set to the given value.

pub fn sequential_dispersed(index: u32) -> Self

Generate a deterministic but quasi-randomly distributed color from a provided index.

This can be helpful for generating debug colors.

Examples

// Unique color for an entity
// let entity_index = entity.index();
let color = Hsla::sequential_dispersed(entity_index);

// Palette with 5 distinct hues
let palette = (0..5).map(Hsla::sequential_dispersed).collect::<Vec<_>>();